code refactoring: created seperate constant file for reports
